+++ spec/configurator_spec.rb
@@ -31,33 +31,44 @@ RSpec.describe Metalware::Configurator do
   }
 
   let :configurator {
+    make_configurator
+  }
+
+  def make_configurator(hl = highline)
     Metalware::Configurator.new(
-      highline: highline,
+      highline: hl,
       configure_file: configure_file_path,
       questions_section: 'test',
       answers_file: answers_file_path
     )
-  }
+  end
 
   def define_questions(questions_hash)
     configure_file.write(questions_hash.to_yaml)
     configure_file.rewind
   end
 
-  def configure_with_input(input_string)
-    $stdout = Tempfile.open
-
-    input.write(input_string)
-    input.rewind
-
-    configurator.configure
-    $stdout.close
+  def redirect_stdout(&block)
+    $stdout = tmp = Tempfile.new
+    block.call
+    tmp.close
   rescue => e
-    $stdout.rewind
-    STDERR.puts $stdout.read
+    begin
+      $stdout.rewind
+      STDERR.puts $stdout.read
+    rescue
+    end
     raise e
   ensure
     $stdout = STDOUT
+  end
+
+  def configure_with_input(input_string)
+    redirect_stdout do
+      input.write(input_string)
+      input.rewind
+      configurator.configure
+    end
   end
 
   describe '#configure' do
@@ -251,7 +262,6 @@ RSpec.describe Metalware::Configurator do
 
     it 'loads default values' do
 
-
       str_ans = "I am a little teapot!!"
       erb_ans = '<%= I_am_an_erb_tag %>'
 
@@ -281,6 +291,39 @@ RSpec.describe Metalware::Configurator do
         'string_erb' => erb_ans,
         'integer_q' => 10
       })
+    end
+
+    it 'loads the old answers as defaults' do
+      define_questions({
+        test: {
+          string_q: {
+            question: "String?",
+            default: "This is the wrong string"
+          },
+          integer_q: {
+            question: 'Integer?',
+            type: 'integer',
+            default: 10
+          }
+        }
+      })
+
+      old_answers = {
+        "string_q" => "CORRECT",
+        "integer_q" => -100,
+        "should_not_see_me" => "OHHH SNAP"
+      }
+      new_answers = old_answers.dup.tap { |h| h.delete("should_not_see_me") }
+
+      first_run_configure = nil
+      redirect_stdout do
+        first_run_configure = make_configurator(HighLine.new)
+        first_run_configure.send(:save_answers, old_answers)
+      end
+      expect(first_run_configure.send(:old_answers)).to eq(old_answers)
+
+      configure_with_input("\n\n\n\n\n\n")
+      expect(answers).to eq(new_answers)
     end
   end
 end